Reorganização dos arquivos de teste.
